% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{robo_care}
\alias{robo_care}
\title{A dataset containing survey results on the acceptance of robot care}
\format{An object of class \code{"tibble"}
\describe{
\item{age}{Age of the participant}
\item{gender}{Gender of the participant}
\item{job_type}{Occupational status.}
\item{robo_bed/human_bed}{How willing are you to accept that a robot will bring you to bed (disagree 1-6 agree).}
\item{robo_XXX/human_XXX}{Several measurements about care (being fed, giving medicine, body massage, washing hair, washing the face, lifted onto a toilet)}
\item{diff_pref}{Preference for difficult tasks.}
\item{technical_knowledge}{How well a participant subjectively feels informed about technical things.}
\item{privacy_concerns}{How concerned the person is about privacy.}
\item{cse}{Self-efficacy with regards to computerized systems. Beier (1999)}
\item{automation_tendency}{The tendency to favor automation over manually doing things.}
\item{care_experience}{Has the participant made direct or indirect experience in caring for a person.}
}}
\source{
Data was self-collected at the RWTH Aachen University.
}
\usage{
data(robo_care)
}
\description{
This data set contains the results of a survey from a study on the acceptance of robot care. The data set contains
some demographic data, some psychometric measurements, as well as attitude measurements towards both human and robot care.
The sample is not representative.
}
\examples{
library(ggplot2)
\donttest{ggplot(robo_care) + aes(cse, robot_care_pref) +
       geom_jitter(width = 0.1, height = 0, alpha = 0.5)}
}
\references{
André Calero Valdez (Eds.) (2017) Akzeptanz autonomer Robotik, Apprimus Aachen
(\href{https://www.apprimus-verlag.de/akzeptanz-autonomer-robotik.html}{Apprimus})
}
\keyword{datasets}
