% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{robo_care_raw}
\alias{robo_care_raw}
\title{A raw data set that can be used in training for factor analysis}
\format{
A data frame with 337 rows and 61 labelled variables:
\describe{
\item{age}{Age of the participant}
\item{gender}{Gender of the participant}
}
}
\source{
Data was self-collected at the RWTH Aachen University.
}
\usage{
data(robo_care_raw)
}
\description{
The dataset contains variable labels and level labels. Use of the
\code{to_factor} function from the \code{labelled} package is highly suggested.
}
\references{
André Calero Valdez (Eds.) (2017) Akzeptanz autonomer Robotik, Apprimus Aachen
(\href{https://www.apprimus-verlag.de/akzeptanz-autonomer-robotik.html}{Apprimus})
}
\keyword{datasets}
