% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{simpson_paradox}
\alias{simpson_paradox}
\title{A data set to demonstrate the simpson paradox}
\format{
A data frame with 2000 rows and 3 variables:
\describe{
\item{V1}{First Variable}
\item{V2}{Second Variable}
\item{group}{Grouping variable}
}
}
\source{
Data was generated.
}
\usage{
simpson_paradox
}
\description{
A data set to demonstrate the simpson paradox
}
\examples{
library(ggplot2)
\donttest{ggplot(simpson_paradox ) + aes(x = V2, y = V1) +
       geom_point()}
}
\keyword{datasets}
